+++ src/cli/shared.js
@@ -1,0 +1,5 @@
+// @ts-check
+
+export const env = {
+  DEBUG: process.env.DEBUG !== undefined && process.env.DEBUG !== '0',
+}
